+++ backend/canisters/community/impl/src/updates/end_video_call.rs
@@ -16,7 +16,11 @@ fn end_video_call(args: Args) -> Response {
 
 fn end_video_call_impl(args: Args, state: &mut RuntimeState) -> Response {
     if let Some(channel) = state.data.channels.get_mut(&args.channel_id) {
-        match channel.chat.events.end_video_call(args.message_id.into(), state.env.now()) {
+        match channel.chat.events.end_video_call(
+            args.message_id.into(),
+            state.env.now(),
+            Some(&mut state.data.event_sink_client),
+        ) {
             EndVideoCallResult::Success => {
                 handle_activity_notification(state);
                 Success