+++ backend/canisters/user/impl/src/updates/end_video_call.rs
@@ -1,7 +1,7 @@
 use crate::guards::caller_is_video_call_operator;
 use crate::{mutate_state, run_regular_jobs, RuntimeState};
 use canister_tracing_macros::trace;
-use chat_events::EndVideoCallResult;
+use chat_events::{EndVideoCallResult, Reader};
 use ic_cdk_macros::update;
 use user_canister::end_video_call::{Response::*, *};
 
@@ -15,7 +15,18 @@ fn end_video_call(args: Args) -> Response {
 
 fn end_video_call_impl(args: Args, state: &mut RuntimeState) -> Response {
     if let Some(chat) = state.data.direct_chats.get_mut(&args.user_id.into()) {
-        match chat.events.end_video_call(args.message_id.into(), state.env.now()) {
+        let was_started_by_me = chat
+            .events
+            .main_events_reader()
+            .message_internal(args.message_id.into())
+            .map(|m| m.sender != args.user_id)
+            .unwrap_or_default();
+
+        match chat.events.end_video_call(
+            args.message_id.into(),
+            state.env.now(),
+            was_started_by_me.then_some(&mut state.data.event_sink_client),
+        ) {
             EndVideoCallResult::Success => Success,
             EndVideoCallResult::MessageNotFound => MessageNotFound,
             EndVideoCallResult::AlreadyEnded => AlreadyEnded,